[Update] Add comments to the code

+++ internet.js
@@ -38,23 +38,39 @@ module.exports = (app) => {
         }
     })
 
+    /*
+     * Displays policy of AMNet services
+     */
     app.get('/internet/policy/', (req, res) => {
         res.render('internet/policy.html.twig', {data: req.session});
     });
 
+    /*
+     * Delete a request created by user
+     */
     app.get('/internet/delete-access/', (req, res) => {
-        connection.query('SELECT * FROM access WHERE access_id = ? AND access_user = ?', [req.query.access_id, req.query.user_id], function(errors, results, fields){
-            if(results.length > 0){
-                connection.query('INSERT INTO admin_actions(action_type, action_user, action_data) VALUES("delete-access", ?, ?)', [req.query.user_id, results[0]['access_mac']]);
-                connection.query('DELETE FROM access WHERE access_id = ?', [req.query.access_id]);
-                res.redirect('/internet/list-access');
-            }
-            else {
-                res.redirect('/access/login/');
-            }
-        });
+        if(!req.session['logged_in']){
+            req.session.returnTo = '/internet/admin-access/';
+            res.redirect('/access/login/');
+        }
+        else {
+            connection.query('SELECT * FROM access WHERE access_id = ? AND access_user = ?', [req.query.access_id, req.query.user_id], function(errors, results, fields){
+                if(results.length > 0){
+                    connection.query('INSERT INTO admin_actions(action_type, action_user, action_data) VALUES("delete-access", ?, ?)', [req.query.user_id, results[0]['access_mac']]);
+                    connection.query('DELETE FROM access WHERE access_id = ?', [req.query.access_id]);
+                    res.redirect('/internet/list-access');
+                }
+                else {
+                    res.redirect('/access/login/');
+                }
+            });
+        }
     });
 
+    /*
+     * Displays a form to allow user to register
+     * new MAC address
+     */
     app.get('/internet/access-request/', (req, res) => {
         if(!req.session['logged_in']){
             req.session.returnTo = '/internet/access-request/';
@@ -65,6 +81,9 @@ module.exports = (app) => {
         }
     });
 
+    /*
+     * Handle request adding, gathering data from access-request form
+     */
     app.post('/internet/add-request/', urlencodedParser, (req, res) => {
         let mac_addr = req.body.mac_addr.replace(/-/g, ':');
         let description = req.body.description;
@@ -76,6 +95,10 @@ module.exports = (app) => {
         });
     });
 
+    /*
+     * Displays all the currently saved requests and display them
+     * in tables to allow easier management from administrators
+     */
     app.get('/internet/admin-access/', (req, res) => {
         if(!req.session['logged_in']){
             req.session.returnTo = '/internet/admin-access/';
@@ -107,6 +130,9 @@ module.exports = (app) => {
         }
     });
 
+    /*
+     * Allows an administrator to mark a request as "agreed"
+     */
     app.get('/internet/allow/:access_id', (req, res) => {
         if(!req.session['logged_in']){
             req.session.returnTo = '/internet/admin-access/';
@@ -126,6 +152,9 @@ module.exports = (app) => {
         }
     });
 
+    /*
+     * Allows an administrator to mark a request as "declined"
+     */
     app.get('/internet/disallow/:access_id', (req, res) => {
         if(!req.session['logged_in']){
             req.session.returnTo = '/internet/admin-access/';
@@ -145,6 +174,10 @@ module.exports = (app) => {
         }
     });
 
+    /*
+     * Delete a request created by user
+     * Only possible for administrators
+     */
     app.get('/internet/delete/:access_id', (req, res) => {
         if(!req.session['logged_in']){
             req.session.returnTo = '/internet/admin-access/';
